Move integration faqs to template page folder
